Refactor failedResponse and FilingResponseUtils
