+++ check-digit/src/main/scala/hmda/uli/api/http/ULIHttpApi.scala
@@ -21,6 +21,7 @@ import hmda.uli.validation.ULI._
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
 import io.circe.generic.auto._
 import hmda.api.http.codec.ErrorResponseCodec._
+import hmda.util.http.FilingResponseUtils.failedResponse
 
 import scala.concurrent.ExecutionContext
 import hmda.util.streams.FlowUtils._
@@ -47,10 +48,7 @@ trait ULIHttpApi extends HmdaTimeDirectives {
                   val uli = ULI(loan.loanId, digit, loan.loanId + digit)
                   complete(ToResponseMarshallable(uli))
                 case Failure(error) =>
-                  val errorResponse =
-                    ErrorResponse(400, error.getLocalizedMessage, uri.path)
-                  complete(ToResponseMarshallable(
-                    StatusCodes.BadRequest -> errorResponse))
+                  failedResponse(StatusCodes.BadRequest, uri, error)
               }
             } ~
               fileUpload("file") {
@@ -108,10 +106,7 @@ trait ULIHttpApi extends HmdaTimeDirectives {
                     val validated = ULIValidated(value)
                     complete(ToResponseMarshallable(validated))
                   case Failure(error) =>
-                    val errorResponse =
-                      ErrorResponse(400, error.getLocalizedMessage, uri.path)
-                    complete(ToResponseMarshallable(
-                      StatusCodes.BadRequest -> errorResponse))
+                    failedResponse(StatusCodes.BadRequest, uri, error)
                 }
               } ~
                 fileUpload("file") {
