+++ hmda/src/test/scala/hmda/api/http/filing/submissions/UploadHttpApiSpec.scala
@@ -10,6 +10,7 @@ import org.scalatest.MustMatchers
 import akka.actor.typed.scaladsl.adapter._
 import akka.cluster.typed.{Cluster, Join}
 import akka.http.scaladsl.model.StatusCodes
+import akka.http.scaladsl.model.Uri.Path
 import akka.kafka.ConsumerSettings
 import akka.stream.scaladsl.Source
 import hmda.api.http.filing.FileUploadUtils
@@ -38,6 +39,8 @@ import hmda.model.institution.Institution
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
 import io.circe.generic.auto._
 import hmda.api.http.codec.filing.submission.SubmissionStatusCodec._
+import hmda.api.http.model.ErrorResponse
+import hmda.api.http.codec.ErrorResponseCodec._
 import hmda.model.institution.InstitutionGenerators.institutionGen
 import hmda.model.submission.SubmissionGenerator.submissionGen
 import hmda.persistence.AkkaCassandraPersistenceSpec
@@ -145,6 +148,9 @@ class UploadHttpApiSpec
   val url =
     s"/institutions/${sampleInstitution.LEI}/filings/$period/submissions/1"
 
+  val badUrl =
+    s"/institutions/${sampleInstitution.LEI}/filings/$period/submissions/2"
+
   val ts = tsGen.sample.getOrElse(TransmittalSheet())
   val tsCsv = ts.toCSV + "\n"
   val tsSource = Source.fromIterator(() => List(tsCsv).iterator)
@@ -172,6 +178,15 @@ class UploadHttpApiSpec
         submission.start must be < System.currentTimeMillis()
         submission.id mustBe submissionId
         submission.status mustBe Uploaded
+      }
+    }
+
+    "return Bad Request when submission doesn't exist" in {
+      Post(badUrl, hmdaFile) ~> uploadRoutes ~> check {
+        status mustBe StatusCodes.BadRequest
+        val response = responseAs[ErrorResponse]
+        response.path mustBe Path(badUrl)
+        response.message mustBe s"Submission ${sampleInstitution.LEI}-${period}-2 not available for upload"
       }
     }
   }
